ZES: Refactor existing content
- use named exports
- move maps from components to page-content-folder
- add types in a more standard way
