--- src/pages/ZESPlus-Research/components/Logos/index.ts
@@ -1,3 +1,0 @@
-import Logos from './Logos';
-
-export default Logos;
