--- src/pages/ZESPlus-Research/components/MapNetworks/index.ts
@@ -1,3 +1,0 @@
-import MapNetworks from './MapNetworks';
-
-export default MapNetworks;
