--- src/pages/ZESPlus-Research/components/MapPlaceholder/index.ts
@@ -1,3 +1,0 @@
-import MapPlaceholder from './MapPlaceholder';
-
-export default MapPlaceholder;
