--- src/pages/ZESPlus-Research/components/MapQuality/index.ts
@@ -1,3 +1,0 @@
-import MapQuality from './MapQuality';
-
-export default MapQuality;
