--- src/pages/ZESPlus-Research/components/MapSafety/index.ts
@@ -1,3 +1,0 @@
-import MapSafety from './MapSafety';
-
-export default MapSafety;
