+++ src/pages/ZESPlus-Research/forschungsprojekt/02_data.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { Heading, List, Paragraph, SectionProps } from '~/components2/Article';
 import { AnchorButton } from '~/components2/Button';
 import { Link } from '~/components2/Link';
-import ButtonWrapper from '../components/ButtonWrapper';
+import { ButtonWrapper } from '../components/ButtonWrapper';
 
 const StyledList = styled(List)`
   list-style-type: lower-alpha;
